+++ sdk/python/pulumi_vsphere/vmfs_datastore.py
@@ -14,7 +14,7 @@ class VmfsDatastore(pulumi.CustomResource):
     The connectivity status of the datastore. If this is `false`,
     some other computed attributes may be out of date.
     """
-    capacity: pulumi.Output[int]
+    capacity: pulumi.Output[float]
     """
     Maximum capacity of the datastore, in megabytes.
     """
@@ -45,7 +45,7 @@ class VmfsDatastore(pulumi.CustomResource):
     `/dc1/datastore/foo/bar/terraform-test`. Conflicts with
     `datastore_cluster_id`.
     """
-    free_space: pulumi.Output[int]
+    free_space: pulumi.Output[float]
     """
     Available space of this datastore, in megabytes.
     """
@@ -76,7 +76,7 @@ class VmfsDatastore(pulumi.CustomResource):
     The IDs of any tags to attach to this resource. See
     [here][docs-applying-tags] for a reference on how to apply tags.
     """
-    uncommitted_space: pulumi.Output[int]
+    uncommitted_space: pulumi.Output[float]
     """
     Total additional storage space, in megabytes,
     potentially used by all virtual machines on this datastore.
